bookmark/requirements.txt
django==5.0
djangorestframework==3.14.0
djoser==2.2.2
python-dotenv==1.0.0
drf-spectacular==0.26.5
psycopg2-binary==2.9.9
gunicorn==21.2.0
beautifulsoup4==4.12.2
lxml==4.9.3